big refactor to use proper absolute transform calculations
