+++ package.json
@@ -27,7 +27,7 @@
     "@storybook/react": "^8.2.5",
     "@storybook/react-vite": "^8.2.5",
     "@storybook/test": "^8.2.5",
-    "@tscircuit/core": "^0.0.130",
+    "@tscircuit/core": "^0.0.147",
     "@tscircuit/plop": "^0.0.10",
     "@types/bun": "^1.1.9",
     "bun-match-svg": "^0.0.6",
